unstage phal f4

+++ firmware/common/phal_F4/adc/adc.c
@@ -10,7 +10,7 @@
 
 bool PHAL_initADC(ADC_TypeDef* adc, ADCInitConfig_t* config, ADCChannelConfig_t channels[], uint8_t num_channels) {
     // Enable clock to the selected peripheral
-    RCC->APB2ENR |= (1U << (RCC_APB2ENR_ADC1EN_Pos + config->adc_number - 1)) & (0x7UL << RCC_APB2ENR_ADC1EN_Pos);
+    RCC->APB2ENR |= (1 << (RCC_APB2ENR_ADC1EN_Pos + config->adc_number - 1)) & (0x7UL << RCC_APB2ENR_ADC1EN_Pos);
 
     // Set prescaler (todo maintain acceptable bounds)
     ADC123_COMMON->CCR &= ~(ADC_CCR_ADCPRE_Msk);
@@ -32,12 +32,12 @@ bool PHAL_initADC(ADC_TypeDef* adc, ADCInitConfig_t* config, ADCChannelConfig_t 
 
     // Regular channel sequence length
     adc->SQR1 &= ~(ADC_SQR1_L);
-    adc->SQR1 |= ((uint32_t)(num_channels - 1) << ADC_SQR1_L_Pos) & ADC_SQR1_L_Msk;
+    adc->SQR1 |= ((num_channels - 1) << ADC_SQR1_L_Pos) & ADC_SQR1_L_Msk;
 
     // DMA configuration
     if (config->dma_mode != ADC_DMA_OFF) {
         // Circular or one shot
-        adc->CR2 |= (ADC_CR2_DMA) | (((uint32_t)(config->dma_mode == ADC_DMA_CIRCULAR) << ADC_CR2_DDS_Pos) & ADC_CR2_DDS_Msk);
+        adc->CR2 |= (ADC_CR2_DMA) | (((config->dma_mode == ADC_DMA_CIRCULAR) << ADC_CR2_DDS_Pos) & ADC_CR2_DDS_Msk);
     } else {
         // Disable ADC DMA Mode
         adc->CR2 &= ~(ADC_CR2_DMA);
